refactor(parity): split expectations into per-domain modules
