+++ tests/stories/harness/parity/expectations.ts
@@ -4,265 +4,33 @@
 // See LICENSE in the project root for details.
 
 /**
- * Declares each provider parity group ONCE as operations-plus-assertions over
- * the `TaskProvider` interface, comparing canonicalized outputs. The fake-binding
- * test (`expectations.fake.test.ts`) runs every group against `MemoryTaskProvider`
- * in this frozen story-contracts lane. `tests/e2e/parity/` imports `PARITY_GROUPS`
- * OUTWARD to run the same groups against real Kaneo — the exported names, types,
- * and each group's `id`/`title` are a hard contract for that lane and for the
- * ledger that mints `@1` catalog storyIds from them.
+ * Aggregates the per-domain parity groups into one ordered PARITY_GROUPS.
+ * Each domain module (./expectations/*.ts) declares its groups once as
+ * operations-plus-assertions over the TaskProvider interface; the fake binding
+ * (expectations.fake.test.ts) and tests/e2e/parity/ both iterate PARITY_GROUPS,
+ * and the story catalog mints one @1 id per group's id/title. This file stays a
+ * value module (it builds PARITY_GROUPS and PARITY_EXCLUSIONS) — not a re-export
+ * barrel — so oxc/no-barrel-file does not fire.
  */
 
-import { expect } from 'bun:test'
+import { commentGroups } from './expectations/comments.js'
+import { errorGroups } from './expectations/errors.js'
+import { projectGroups } from './expectations/projects.js'
+import { relationGroups } from './expectations/relations.js'
+import { searchGroups } from './expectations/search.js'
+import { taskGroups } from './expectations/tasks.js'
+import type { ParityGroup } from './group.js'
 
-import type { TaskProvider } from '../../../../src/providers/types.js'
-import { canonicalize, VOLATILE, VOLATILE_KEYS } from './canonicalize.js'
-
-export type ParityHarness = Readonly<{
-  provider: TaskProvider
-  projectId: string
-}>
-
-export type ParityGroup = Readonly<{
-  id: string
-  title: string
-  run(harness: ParityHarness): Promise<void>
-}>
-
-/** Unwraps an optional-method result. Both `MemoryTaskProvider` and `KaneoProvider`
- *  implement every method a parity group calls, so this never throws at runtime —
- *  it only satisfies the `TaskProvider` optional-method type without `!`. */
-function required<T>(value: T | undefined, label: string): T {
-  if (value === undefined) throw new Error(`parity: expected ${label} to be defined`)
-  return value
-}
+export type { ParityGroup, ParityHarness } from './group.js'
 
 export const PARITY_GROUPS: readonly ParityGroup[] = [
-  {
-    id: 'SCN-parity-task-create',
-    title: 'SCN-parity-task-create: createTask returns a normalized task shape',
-    async run({ provider, projectId }) {
-      const task = await provider.createTask({ projectId, title: 'Parity Create' })
-      expect(canonicalize(task, VOLATILE_KEYS)).toMatchObject({
-        id: VOLATILE,
-        projectId: VOLATILE,
-        title: 'Parity Create',
-      })
-      const createdUrl = required(task.url, 'task.url')
-      expect(createdUrl).toBeTypeOf('string')
-      expect(createdUrl.length).toBeGreaterThan(0)
-    },
-  },
-  {
-    id: 'SCN-parity-task-get',
-    title: 'SCN-parity-task-get: getTask returns the same normalized shape as createTask',
-    async run({ provider, projectId }) {
-      const created = await provider.createTask({ projectId, title: 'Parity Get' })
-      const fetched = await provider.getTask(created.id)
-      expect(canonicalize(fetched, VOLATILE_KEYS)).toMatchObject({
-        id: VOLATILE,
-        projectId: VOLATILE,
-        title: 'Parity Get',
-      })
-      const fetchedUrl = required(fetched.url, 'fetched.url')
-      expect(fetchedUrl).toBeTypeOf('string')
-      expect(fetchedUrl.length).toBeGreaterThan(0)
-    },
-  },
-  {
-    id: 'SCN-parity-task-update',
-    title: 'SCN-parity-task-update: updateTask applies a title and status change',
-    async run({ provider, projectId }) {
-      const created = await provider.createTask({ projectId, title: 'Parity Update' })
-      const updated = await provider.updateTask(created.id, { title: 'Parity Updated', status: 'In Progress' })
-      // `status` is not pinned to the literal column name here: real Kaneo echoes a
-      // normalized/slugged form (e.g. "in-progress") rather than "In Progress", while
-      // the fake echoes the input verbatim — both are non-empty strings, so require
-      // only presence and type, not a fixed value.
-      expect(canonicalize(updated, VOLATILE_KEYS)).toMatchObject({
-        id: VOLATILE,
-        projectId: VOLATILE,
-        title: 'Parity Updated',
-      })
-      const updatedStatus = required(updated.status, 'updated.status')
-      expect(updatedStatus).toBeTypeOf('string')
-      expect(updatedStatus.length).toBeGreaterThan(0)
-      const updatedUrl = required(updated.url, 'updated.url')
-      expect(updatedUrl).toBeTypeOf('string')
-      expect(updatedUrl.length).toBeGreaterThan(0)
-      const fetched = await provider.getTask(created.id)
-      expect(canonicalize(fetched, VOLATILE_KEYS)).toMatchObject({
-        id: VOLATILE,
-        projectId: VOLATILE,
-        title: 'Parity Updated',
-      })
-      const fetchedUrl = required(fetched.url, 'fetched.url')
-      expect(fetchedUrl).toBeTypeOf('string')
-      expect(fetchedUrl.length).toBeGreaterThan(0)
-      const fetchedStatus = required(fetched.status, 'fetched.status')
-      expect(fetchedStatus).toBeTypeOf('string')
-      expect(fetchedStatus).toBe(updatedStatus)
-    },
-  },
-  {
-    id: 'SCN-parity-task-delete',
-    title: 'SCN-parity-task-delete: deleteTask removes the task and getTask then rejects',
-    async run({ provider, projectId }) {
-      const created = await provider.createTask({ projectId, title: 'Parity Delete' })
-      const deleted = await provider.deleteTask?.(created.id)
-      expect(canonicalize(deleted, VOLATILE_KEYS)).toEqual({ id: VOLATILE })
-      await expect(provider.getTask(created.id)).rejects.toThrow()
-    },
-  },
-  {
-    id: 'SCN-parity-task-list-sort',
-    title: 'SCN-parity-task-list-sort: listTasks honors sortBy/sortOrder across providers',
-    async run({ provider, projectId }) {
-      // Seed in a deliberately non-sorted creation order so the assertion below can
-      // only pass if `sortBy`/`sortOrder` actually reordered the results.
-      await provider.createTask({ projectId, title: 'Sort B' })
-      await provider.createTask({ projectId, title: 'Sort C' })
-      await provider.createTask({ projectId, title: 'Sort A' })
-      const listed = await provider.listTasks(projectId, { sortBy: 'title', sortOrder: 'asc' })
-      const titles = listed.map((task) => task.title)
-      expect(titles).toEqual(['Sort A', 'Sort B', 'Sort C'])
-      expect(canonicalize(listed, VOLATILE_KEYS)).toHaveLength(3)
-    },
-  },
-  {
-    id: 'SCN-parity-task-list-paging',
-    title: 'SCN-parity-task-list-paging: listTasks pages seeded tasks with limit and page',
-    async run({ provider, projectId }) {
-      await provider.createTask({ projectId, title: 'Page A' })
-      await provider.createTask({ projectId, title: 'Page B' })
-      await provider.createTask({ projectId, title: 'Page C' })
-      const firstPage = await provider.listTasks(projectId, { limit: 2, page: 1 })
-      const secondPage = await provider.listTasks(projectId, { limit: 2, page: 2 })
-      expect(firstPage.map((task) => task.title)).toEqual(['Page A', 'Page B'])
-      expect(secondPage.map((task) => task.title)).toEqual(['Page C'])
-    },
-  },
-  {
-    id: 'SCN-parity-task-search',
-    title: 'SCN-parity-task-search: searchTasks matches seeded tasks by query',
-    async run({ provider, projectId }) {
-      await provider.createTask({ projectId, title: 'Searchable Falcon' })
-      await provider.createTask({ projectId, title: 'Searchable Osprey' })
-      await provider.createTask({ projectId, title: 'Unrelated Item' })
-      const results = await provider.searchTasks({ query: 'Searchable', projectId })
-      const titles = results.map((result) => result.title).sort()
-      expect(titles).toEqual(['Searchable Falcon', 'Searchable Osprey'])
-    },
-  },
-  {
-    id: 'SCN-parity-comment-crud',
-    title: 'SCN-parity-comment-crud: add, list, update, and remove a comment',
-    async run({ provider, projectId }) {
-      const task = await provider.createTask({ projectId, title: 'Comment Host' })
-      const added = await provider.addComment?.(task.id, 'first note')
-      expect(canonicalize(added, VOLATILE_KEYS)).toMatchObject({ id: VOLATILE, body: 'first note' })
-      const listed = (await provider.getComments?.(task.id, {})) ?? []
-      expect(listed.map((comment) => comment.body).sort()).toEqual(['first note'])
-      const addedId = required(added, 'addComment result').id
-      const updated = await provider.updateComment?.({ taskId: task.id, commentId: addedId, body: 'edited note' })
-      expect(canonicalize(updated, VOLATILE_KEYS)).toMatchObject({ id: VOLATILE, body: 'edited note' })
-      const removed = await provider.removeComment?.({ taskId: task.id, commentId: addedId })
-      expect(canonicalize(removed, VOLATILE_KEYS)).toEqual({ id: VOLATILE })
-    },
-  },
-  {
-    id: 'SCN-parity-task-label',
-    title: 'SCN-parity-task-label: attach and detach a label from a task',
-    async run({ provider, projectId }) {
-      const task = await provider.createTask({ projectId, title: 'Label Host' })
-      const label = await provider.createLabel?.({ name: 'attach-label' })
-      const labelId = required(label, 'createLabel result').id
-      const attached = await provider.addTaskLabel?.(task.id, labelId)
-      expect(canonicalize(attached, VOLATILE_KEYS)).toEqual({ taskId: VOLATILE, labelId: VOLATILE })
-      const listed = (await provider.listTaskLabels?.(task.id)) ?? []
-      expect(listed.map((entry) => entry.name)).toEqual(['attach-label'])
-      const removed = await provider.removeTaskLabel?.(task.id, labelId)
-      expect(canonicalize(removed, VOLATILE_KEYS)).toEqual({ taskId: VOLATILE, labelId: VOLATILE })
-    },
-  },
-  {
-    id: 'SCN-parity-project-crud',
-    title: 'SCN-parity-project-crud: create, list, update, and delete a project',
-    async run({ provider }) {
-      const created = required(await provider.createProject?.({ name: 'Parity Project CRUD' }), 'createProject result')
-      expect(Object.keys(created).sort()).toEqual(['id', 'name', 'url'])
-      expect(canonicalize(created, VOLATILE_KEYS)).toMatchObject({ id: VOLATILE, name: 'Parity Project CRUD' })
-      const listed = (await provider.listProjects?.()) ?? []
-      expect(listed.map((project) => project.name)).toContain('Parity Project CRUD')
-      const updated = required(
-        await provider.updateProject?.(created.id, { name: 'Parity Project Renamed' }),
-        'updateProject result',
-      )
-      expect(Object.keys(updated).sort()).toEqual(['id', 'name', 'url'])
-      expect(canonicalize(updated, VOLATILE_KEYS)).toMatchObject({ id: VOLATILE, name: 'Parity Project Renamed' })
-      const removed = await provider.deleteProject?.(created.id)
-      expect(canonicalize(removed, VOLATILE_KEYS)).toEqual({ id: VOLATILE })
-    },
-  },
-  {
-    id: 'SCN-parity-relation',
-    title: 'SCN-parity-relation: add, update, and remove a task relation',
-    async run({ provider, projectId }) {
-      const first = await provider.createTask({ projectId, title: 'Relation First' })
-      const second = await provider.createTask({ projectId, title: 'Relation Second' })
-      const added = await provider.addRelation?.(first.id, second.id, 'blocks')
-      expect(canonicalize(added, VOLATILE_KEYS)).toEqual({ taskId: VOLATILE, relatedTaskId: VOLATILE, type: 'blocks' })
-      const updated = await provider.updateRelation?.(first.id, second.id, 'related')
-      expect(canonicalize(updated, VOLATILE_KEYS)).toEqual({
-        taskId: VOLATILE,
-        relatedTaskId: VOLATILE,
-        type: 'related',
-      })
-      const removed = await provider.removeRelation?.(first.id, second.id)
-      expect(canonicalize(removed, VOLATILE_KEYS)).toEqual({ taskId: VOLATILE, relatedTaskId: VOLATILE })
-    },
-  },
-  {
-    id: 'SCN-parity-identity',
-    title: 'SCN-parity-identity: provisionWorkspaceMember and listUsers resolve normalized shapes',
-    async run({ provider }) {
-      const provisioned = required(
-        await provider.provisionWorkspaceMember?.({
-          chatUserId: 'parity-alice',
-          displayName: 'Parity Alice',
-          username: 'parity.alice',
-        }),
-        'provisionWorkspaceMember result',
-      )
-      expect(Object.keys(provisioned).sort()).toEqual(['login', 'password', 'providerUserId'])
-      // `login`/`password`/`providerUserId` are provider-opaque (not in VOLATILE_KEYS):
-      // each provider mints them differently — Kaneo synthesizes a unique email login
-      // rather than echoing the requested username — so require only presence and type,
-      // not a fixed sentinel or literal.
-      expect(provisioned.login).toBeTypeOf('string')
-      expect(provisioned.login.length).toBeGreaterThan(0)
-      expect(provisioned.password.length).toBeGreaterThan(0)
-      expect(provisioned.providerUserId.length).toBeGreaterThan(0)
-      // The fake's provisionWorkspaceMember doesn't populate the store listUsers reads
-      // from, so element-shape parity can't be asserted hermetically here; the strong
-      // cross-provider identity signal is the provisionWorkspaceMember assertion above.
-      const users = required(await provider.listUsers?.('parity', 10), 'listUsers result')
-      expect(Array.isArray(users)).toBe(true)
-      // When a provider does surface the just-provisioned member (real Kaneo matches it
-      // by the 'parity' substring on name/email; the fake returns [] so this is a no-op
-      // there), assert the normalized UserRef shape. A for...of over the filtered matches
-      // keeps the check tolerant of the empty fake result without a conditional expect.
-      const provisionedMembers = users.filter((user) => user.login === provisioned.login)
-      for (const member of provisionedMembers) {
-        expect(member.id).toBeTypeOf('string')
-        expect(member.id.length).toBeGreaterThan(0)
-        const memberName = required(member.name, 'listUsers member.name')
-        expect(memberName).toBeTypeOf('string')
-        expect(memberName.length).toBeGreaterThan(0)
-      }
-    },
-  },
-] as const
+  ...taskGroups,
+  ...searchGroups,
+  ...commentGroups,
+  ...relationGroups,
+  ...projectGroups,
+  ...errorGroups,
+]
 
 export const PARITY_EXCLUSIONS: readonly Readonly<{ group: string; reason: string }>[] = [
   {
