+++ tests/stories/harness/parity/expectations/errors.ts
@@ -1,0 +1,8 @@
+// SPDX-License-Identifier: BUSL-1.1
+// Copyright (c) 2026 Dmitriy Lazarev
+// Use of this software is governed by the Business Source License 1.1.
+// See LICENSE in the project root for details.
+
+import type { ParityGroup } from '../group.js'
+
+export const errorGroups: readonly ParityGroup[] = [] as const
